Split and rename file
